Switch locals to be based on ContextVars
ContextVars have been introduced into Python 3.7 and are effectively
the same as Locals in Werkzeug (although a different API). They should
work in a Greenlet, and threading, context as before but also now in
an asyncio, or other async/await context.

The __storage__ attribute has been kept for backwards compatibility
(even though it is a dunder it seems it is used, e.g. in tests).

The ident_function though must be deprecated as this is now handled by
the ContextVar. This may cause some backwards incompatibility.

This is based on the Quart local implementation.

If ContextVars aren't available (Python 3.6) or there is an issue
using them i.e. Gevent/Eventlet have no monkey patched them then the
local will fall back to the previous (green) thread identity based
version.